[CONFIG] [Github Actions] C/C++ for windows.

+++ src/lib/exercises/include/exercises/hackerrank/warmup/a_very_big_sum.h
@@ -4,7 +4,7 @@
 extern "C" {
 #endif
 
-long HACKERRANK_WARMUP_aVeryBigSum(int ar_count, const long *ar);
+long long HACKERRANK_WARMUP_aVeryBigSum(int ar_count, const long *ar);
 
 #ifdef __cplusplus
 } // extern "C"
